Fully functional and documented image upload script.
Script can be used by submitting the form in uploadForm.php. Script contents may be moved into uploadForm.php so it won't have to redirect. uploadForm contents may also be moved into index.php later depending on whether the upload form should be a separate site or not.
The script will only be functional with a working SQL database so make sure you go through all the steps mentioned in the README for database creation.